+++ pages/_app.jsx
@@ -2,6 +2,9 @@
 import 'bulma/css/bulma.min.css';
 import '../styles/globals.css';
 import "react-loader-spinner/dist/loader/css/react-spinner-loader.css";
+import { config } from '@fortawesome/fontawesome-svg-core'
+import '@fortawesome/fontawesome-svg-core/styles.css'
+config.autoAddCss = false
 
 import MainLayout from '../layouts/MainLayout';